% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLognormalError.R
\name{addLognormalError}
\alias{addLognormalError}
\title{Add lognormal error}
\usage{
addLognormalError(obs, sigma, biasadjustflag = FALSE, randomval = NULL)
}
\arguments{
\item{obs}{a numer or vector of starting value(s)}

\item{sigma}{lognormal standard deviation to apply, can be single value or vector of same length as obs}

\item{biasadjustflag}{TRUE means to apply bias adjustment, default = FALSE}

\item{randomval}{number of vector of predetermined random values to use, default = NULL}
}
\description{
Multiply an observation by exp(sigma) with optional bias adjustment of -sigma^2/2.
}
