% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{blockReplace}
\alias{blockReplace}
\title{nd a block of text (Find1) and replace with another block (Replace1) in all files (with extension - ext or containing - ext)
 directory (dnm) - this can be done recursively or not. You will by default also create backups
 any files that you alter, but you can turn that off with backup=F
             NOTE: Find1, Replace1 and ext all must be quoted strings!
***** WARNING: Watch out for automatic indentation - disble or use shift+tab to outdent!!!!!}
\usage{
blockReplace(dnm, Find1, Replace1, ext = NULL, recursive = F,
  backup = T)
}
\description{
nd a block of text (Find1) and replace with another block (Replace1) in all files (with extension - ext or containing - ext)
 directory (dnm) - this can be done recursively or not. You will by default also create backups
 any files that you alter, but you can turn that off with backup=F
             NOTE: Find1, Replace1 and ext all must be quoted strings!
***** WARNING: Watch out for automatic indentation - disble or use shift+tab to outdent!!!!!
}
