% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCatch.R
\name{calcCatch}
\alias{calcCatch}
\title{Catch Equation}
\usage{
calcCatch(PopN, Mval, Fval)
}
\arguments{
\item{PopN}{population abundance, can be number, vector, or matrix}

\item{Mval}{natural mortality rate, can be number or same dimension as PopN}

\item{Fval}{fishing mortality rate, can be number or same dimension as PopN}
}
\description{
Compute expected catch given population abundance, natural mortality, and fishing mortality.
}
