% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEquilibriumPop.R
\name{calcEquilibriumPop}
\alias{calcEquilibriumPop}
\title{Equilibrium Population}
\usage{
calcEquilibriumPop(eqR, nages, Mval, Fval, plusgroupflag = TRUE)
}
\arguments{
\item{eqR}{number of recruits in equilibrium population}

\item{nages}{number of ages in population}

\item{Mval}{vector of natural mortality at age or single value applied to all ages}

\item{Fval}{vector of fishing mortality at age or single value applied to all ages}

\item{plusgroupflag}{means last element in vector or matrix is a plus group, default = TRUE}
}
\description{
Computes equilibrium population at age given Recruits, M at age, and F at age.
}
