% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{col2hex}
\alias{col2hex}
\title{convert a color from names "col" to a hexadecimal representation}
\usage{
col2hex(col)
}
\description{
convert a color from names "col" to a hexadecimal representation
}
