% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{error.bar}
\alias{error.bar}
\title{make error bars on a plot
 @param x vector of x values
 @param y vector of y values
 @param upper distance (in units of y) to draw the upper error bar
 @param lower (default=upper)  distance (in units of y) to draw the lower error bar
 @param length length of cross bars at error bar ends
 @param col1 color of error bars
 @param ... other plot parameters}
\usage{
error.bar(x, y, upper, lower = upper, length = 0.1, col1 = "black", ...)
}
\description{
make error bars on a plot
 @param x vector of x values
 @param y vector of y values
 @param upper distance (in units of y) to draw the upper error bar
 @param lower (default=upper)  distance (in units of y) to draw the lower error bar
 @param length length of cross bars at error bar ends
 @param col1 color of error bars
 @param ... other plot parameters
}
