% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{fac_to_num}
\alias{fac_to_num}
\title{Get rid of factors - in this case when you want nums instead}
\usage{
fac_to_num(x)
}
\description{
Get rid of factors - in this case when you want nums instead
}
