% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{filled.contour3}
\alias{filled.contour3}
\title{modification by Ian Taylor of the filled.contour function
to remove the key and facilitate overplotting with contour()
 further modified by Carey McGilliard and Bridget Ferris
 to allow multiple plots on one page}
\usage{
filled.contour3(x = seq(0, 1, length.out = nrow(z)), y = seq(0, 1,
  length.out = ncol(z)), z, xlim = range(x, finite = TRUE),
  ylim = range(y, finite = TRUE), zlim = range(z, finite = TRUE),
  levels = pretty(zlim, nlevels), nlevels = 20,
  color.palette = cm.colors, col = color.palette(length(levels) - 1),
  plot.title, plot.axes, key.title, key.axes, asp = NA, xaxs = "i",
  yaxs = "i", las = 1, axes = TRUE, frame.plot = axes, mar, ...)
}
\description{
modification by Ian Taylor of the filled.contour function
to remove the key and facilitate overplotting with contour()
 further modified by Carey McGilliard and Bridget Ferris
 to allow multiple plots on one page
}
