% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{gamma.ci}
\alias{gamma.ci}
\title{gamma.ci
generate approximate confidence intervals based on gamma variable x}
\usage{
\method{gamma}{ci}(ts, cv, bounds = 95)
}
\arguments{
\item{ts}{is your data}

\item{cv}{its cv}

\item{bounds}{default=95 the width of your confidence intervals}
}
\description{
gamma.ci
generate approximate confidence intervals based on gamma variable x
}
