% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{isdebugged_safe}
\alias{isdebugged_safe}
\title{return the names of the objects (from a vector of list of
 names of objects) that are functions and have debug flag set}
\usage{
isdebugged_safe(x, ns = NULL)
}
\description{
return the names of the objects (from a vector of list of
 names of objects) that are functions and have debug flag set
}
