% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{lineReplace}
\alias{lineReplace}
\title{Find (Find1) and replace with (Replace1) in all files (with extension - or containing - ext)
in directory (dnm) - this can be done recursively or not. You will by default also create backups
of any files that you alter, but you can turn that off with backup=F
              NOTE: Find1, Replace1 and ext all must be quoted strings!}
\usage{
lineReplace(dnm, Find1, Replace1, ext = NULL, recursive = F, backup = T)
}
\arguments{
\item{dnm}{directory to look in}

\item{Find1}{string to find}

\item{Replace1}{string to use as replacement}

\item{ext}{default=NULL an extension, or part of the filename, that will identify the type of file to look in}

\item{recursive}{default=FALSE should the replacement be made to all files in subdirectories below dnm?}

\item{backup}{default=TRUE should a backup of the original be made? This will be stored in a directory called "Backup"
located in the same directory as the file being manipulated.}
}
\description{
Find (Find1) and replace with (Replace1) in all files (with extension - or containing - ext)
in directory (dnm) - this can be done recursively or not. You will by default also create backups
of any files that you alter, but you can turn that off with backup=F
              NOTE: Find1, Replace1 and ext all must be quoted strings!
}
\examples{
dnm="/home/dhennen/Clam/Assessment Stuff/SS3/2016/runs/North/Explorations/TestReplaceFunc/"
Find1=" 3 31 14.1093 10.3 -1 3 1 # SR_LN(R0)"
Replace1=" 3 31 14.1093 10.3 -1 3 -1 # SR_LN(R0)"
ext=".ctl"
lineReplace(dnm,Find1,Replace1,ext)
}
