% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{na.replace}
\alias{na.replace}
\title{This function will replace NA's in
vector x with rep1 (0 is the default)}
\usage{
na.replace(x, rep1 = 0)
}
\arguments{
\item{x}{the vector of interest}

\item{rep1}{the thing you want to swap in for NA (0 is default)}
}
\description{
This function will replace NA's in
vector x with rep1 (0 is the default)
}
