% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{normal.ci}
\alias{normal.ci}
\title{normal.ci
generate approximate confidence intervals based on normal variable x}
\usage{
normal.ci(x, cv.x, bounds = 95)
}
\arguments{
\item{bounds}{default=95 the width of your confidence intervals}

\item{ts}{is your data}

\item{cv}{its cv}
}
\description{
normal.ci
generate approximate confidence intervals based on normal variable x
}
