% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{plot.correlated.status.check}
\alias{plot.correlated.status.check}
\title{Plot and conrast two correlated overlapping lognormal distributions, will make
a plot and return the probability of exceeding the threshold ie overfishing
or overfished status}
\usage{
\method{plot}{correlated.status.check}(mu = NULL, cv1 = NULL,
  cor1 = NULL, aname = NULL, bname = NULL, colors = c("blue",
  "red", "purple"), b1 = 1000, xlim = NULL, ylim = NULL,
  xlab = NULL, ylab = NULL, main = NULL, sub1 = NULL,
  put.legend = "topright", status.test = "overfishing")
}
\arguments{
\item{mu}{default=NULL vector of means of the two distributions one for the estimate, one
for the reference point (given in real space, not logged)}

\item{cv1}{default=NULL vector of cv's}

\item{cor1}{default=NULL correlation of the two above}

\item{aname}{default=NULL  name for the a marginal distribution A[,1] - will go in legend}

\item{bname}{default=NULL  name for the b marginal distribution A[,2] - will go in legend}

\item{xlim}{default=NULL predefined limits (if you want to restrict your plot)}

\item{ylim}{default=NULL}

\item{xlab}{default=NULL other plotting info}

\item{ylab}{default=NULL}

\item{main}{default=NULL}

\item{sub1}{default=NULL  subtext for plot}

\item{colors=c("blue", "red", "purple")}{colors for each marginal distribution and the}

\item{b1=1000}{the number of breaks in the plot of the historgram of a and b}

\item{put.legend="topright"}{where should the legend go?}

\item{status.test="overfishing"}{could be "overfished"}
}
\description{
Plot and conrast two correlated overlapping lognormal distributions, will make
a plot and return the probability of exceeding the threshold ie overfishing
or overfished status
}
\examples{
  plot.correlated.status.check(mu=c(1000,500)
                  ,cv1=c(0.22,0.18)
                  ,cor1=0.9
                  ,aname="B2011"
                  ,bname="Threshold"
                  ,status.test="overfished"
                 )
}
