% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readASAP.R
\name{readASAP}
\alias{readASAP}
\title{Read ASAP as input to OM}
\usage{
readASAP(asap.name, wd)
}
\arguments{
\item{asap.name}{filename without extension of the ASAP run}

\item{wd}{directory where the ASAP run is located}
}
\description{
Uses an ASAP input file and associated rdat file to set up an operating model.
}
