% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{transparent}
\alias{transparent}
\title{Convert a named color to a transparent equivalent}
\usage{
transparent(col, deg)
}
\arguments{
\item{col}{is the color}

\item{deg}{is the amount of transparency desired from 1 (almost invisible)
to 100 (opaque).}
}
\description{
Convert a named color to a transparent equivalent
}
\examples{
transparent("blue",50) #50\% transparent blue
}
