% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{var.ratio}
\alias{var.ratio}
\title{Calculate the variance of ratio x1/x2; where x1 and x2 are vectors of random
variables (x2>0)}
\usage{
var.ratio(x1, x2)
}
\arguments{
\item{x1}{vector of values}

\item{x2}{vector of values (x2>0)}
}
\description{
Calculate the variance of ratio x1/x2; where x1 and x2 are vectors of random
variables (x2>0)
}
