% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{wt.var.ratio}
\alias{wt.var.ratio}
\title{Calculate the variance of ratio x1/x2; where x1 and x2 are vectors of random
variables (x2>0) with weights: w}
\usage{
wt.var.ratio(x1, x2, w)
}
\arguments{
\item{x1}{vector of values}

\item{x2}{vector of values (x2>0)
w vector of weights}
}
\description{
Calculate the variance of ratio x1/x2; where x1 and x2 are vectors of random
variables (x2>0) with weights: w
}
